added extra tests

+++ dtu_pay/dtu_pay_server/src/main/java/dtu/Models/Transaction.java
@@ -1,5 +1,7 @@
 package dtu.Models;
 
-public record Transaction(String customerId, String merchantId, int payment) {
+import java.math.BigDecimal;
+
+public record Transaction(String customerId, String merchantId, BigDecimal payment) {
 
 }
